feat(react): move Split/Unified toggle into FileTree header
Relocate the Split/Unified diff-view ToggleGroup from Toolbar.tsx into
the FileTree header so embedders of @self-review/react retain the
control even when they do not mount the bundled Toolbar. The toggle
adopts the icon-only variant used by the surrounding header buttons
to fit the constrained sidebar width; data-testid selectors and the
useConfig({ diffView }) contract are preserved verbatim.
